+++ src/Contracts/TrackableJobEvent.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Konekt\History\Contracts;
+
+use Konekt\History\JobTracker;
+
+interface TrackableJobEvent
+{
+    public function getJob(): TrackableJob;
+
+    public function getTracker(): JobTracker;
+}