+++ src/Events/TrackableJobCompleted.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Konekt\History\Events;
+
+class TrackableJobCompleted extends BaseTrackableJobEvent
+{
+}